+++ tests/test_application.py
@@ -1,4 +1,4 @@
-from unittest.mock import MagicMock
+from unittest.mock import MagicMock, Mock
 from pytest import fixture, raises
 from limberframework.foundation.application import Application
 
@@ -17,13 +17,15 @@ def test_bind_service(application):
     name = 'test'
     mock_closure = MagicMock()
     singleton = True
+    defer = True
 
-    application.bind(name, mock_closure, singleton)
+    application.bind(name, mock_closure, singleton, defer)
 
     assert application.bindings == {
         name: {
             'closure': mock_closure,
-            'singleton': singleton
+            'singleton': singleton,
+            'defer': defer
         }
     }
 
@@ -60,6 +62,33 @@ def test_make_known_singleton_service(application):
     assert isinstance(service_1, MagicMock)
     assert isinstance(service_2, MagicMock)
     assert service_1 is service_2
+
+def test_load_services(application):
+    mock_closure = Mock()
+    mock_make = Mock()
+
+    application.make = mock_make
+    services = [
+        {
+            'name': 'cache',
+            'closure': mock_closure,
+            'singleton': True,
+            'defer': False
+        },
+        {
+            'name': 'auth',
+            'closure': mock_closure,
+            'singleton': False,
+            'defer': True
+        }
+    ]
+
+    for service in services:
+        application.bind(service['name'], service['closure'], service['singleton'], service['defer'])
+
+    application.load_services()
+
+    application.make.assert_called_once_with(services[0]['name'])
 
 def test_get_item(application):
     name = 'test'